+++ ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/domain/CCDClaim.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.cmc.ccd.domain;
 
 import lombok.Builder;
 import lombok.Value;
+import uk.gov.hmcts.cmc.ccd.domain.evidence.CCDEvidence;
 
 import java.math.BigInteger;
 import java.util.List;
