+++ ccd-adapter/src/main/java/uk/gov/hmcts/cmc/ccd/domain/evidence/CCDEvidence.java
@@ -1,7 +1,8 @@
-package uk.gov.hmcts.cmc.ccd.domain;
+package uk.gov.hmcts.cmc.ccd.domain.evidence;
 
 import lombok.Builder;
 import lombok.Value;
+import uk.gov.hmcts.cmc.ccd.domain.CCDCollectionElement;
 
 import java.util.List;
