+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/evidence/EvidenceRow.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.cmc.domain.models;
+package uk.gov.hmcts.cmc.domain.models.evidence;
 
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
