+++ domain-sample-data/src/main/java/uk/gov/hmcts/cmc/domain/models/sampledata/SampleEvidence.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.cmc.domain.models.sampledata;
 
-import uk.gov.hmcts.cmc.domain.models.Evidence;
-import uk.gov.hmcts.cmc.domain.models.EvidenceRow;
+import uk.gov.hmcts.cmc.domain.models.evidence.Evidence;
+import uk.gov.hmcts.cmc.domain.models.evidence.EvidenceRow;
 
 import java.util.List;
